pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vid-tools"
version = "0.2.0"
description = "Collection of video and image editing CLI tools"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Ben", email = "benl@irontowerpress.com" },
]
keywords = ["video", "slow-motion", "loop", "ffmpeg", "frame-interpolation", "background-video", "runway", "ai-video"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Video :: Conversion",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "opencv-python>=4.8",
    "numpy>=1.24",
    "rembg[cpu]>=2.0",
    "Pillow>=10.0",
    "segno>=1.6.6",
]

[project.scripts]
video-slowmo-loop = "video_slowmo_loop:main"
video-looper = "video_looper:main"

[project.urls]
Homepage = "https://github.com/blarson/vid-tools"
Repository = "https://github.com/blarson/vid-tools"

[tool.hatch.build.targets.wheel]
only-include = ["video_slowmo_loop.py", "video_looper.py"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]